Full ability to list attributes from the DAO

+++ service/internal/service/page.go
@@ -2,5 +2,5 @@ package service
 
 // PageInfo represents pagination information returned by the services
 type PageInfo struct {
-	TotalSize int
+	TotalSize uint64
 }
